System walki i dużo poprawek logistycznych jeśli chodzi o zliczanie rzeczy potrzebnych do questa

+++ src/main/java/com/gametest/springprojekt/dto/QuestDto.java
@@ -21,7 +21,7 @@ public class QuestDto {
 
     private String opponentName;
 
-    private int questTime;
+    private long questTime;
 
     private int moneyReward;
